chatapp/requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
bidict==0.22.1
blinker==1.4
certifi==2019.11.28
cffi==1.16.0
channels==3.0.4
chardet==3.0.4
Click==7.0
colorama==0.4.3
constantly==15.1.0
cryptography==41.0.4
daphne==3.0.2
distro==1.4.0
Django==4.2.6
django-cors-headers==4.3.0
django-phonenumber-field==7.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
ecdsa==0.18.0
entrypoints==0.3
fasteners==0.14.1
future==0.18.2
gunicorn==21.2.0
h11==0.14.0
httplib2==0.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
keyring==18.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
monotonic==1.5
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
packaging==23.2
paramiko==2.6.0
pbr==6.0.0
pexpect==4.6.0
phonenumbers==8.13.23
Pillow==7.0.0
protobuf==3.6.1
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==23.3.0
pyRFC3339==1.1
python-decouple==3.8
python-dotenv==1.0.0
python-engineio==4.8.0
python-jose==3.3.0
python-socketio==5.10.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.22.0
rsa==4.9
SecretStorage==2.3.1
service-identity==23.1.0
simple-websocket==1.0.0
six==1.16.0
sqlparse==0.4.4
testresources==2.0.1
twisted==23.8.0
txaio==23.1.1
typing-extensions==4.8.0
uritemplate==4.1.1
urllib3==1.25.8
uvicorn==0.24.0.post1
wadllib==1.3.3
wsproto==1.2.0
zope.interface==6.1
